+++ inst/shiny-examples/ShinyItemAnalysis/ui/uiReliability.R
@@ -170,7 +170,7 @@ uiReliability <-
       h4("Histogram of reliability estimates"),
       p("Histogram is based on selected number of split halves estimates (10,000 by default).
                         The current estimate is highlighted by red colour."),
-      plotOutput("reliability_splithalf_histogram"),
+      plotlyOutput("reliability_splithalf_histogram"),
       downloadButton("DB_reliability_splithalf_histogram"),
       br(),
       h4("Selected R code"),
@@ -198,6 +198,96 @@ uiReliability <-
       # br(),
       h4("Selected R code"),
       div(code(HTML("library(psychometric)<br>library(ShinyItemAnalysis)<br><br>#&nbsp;loading&nbsp;data<br>data(HCI)<br>data&nbsp;<-&nbsp;HCI[,&nbsp;1:20]<br><br>#&nbsp;Cronbach's&nbsp;alpha&nbsp;with&nbsp;confidence&nbsp;interval<br>a&nbsp;<-&nbsp;psychometric::alpha(data)<br>psychometric::alpha.CI(a,&nbsp;N&nbsp;=&nbsp;nrow(data),&nbsp;k&nbsp;=&nbsp;ncol(data),&nbsp;level&nbsp;=&nbsp;0.95)"))),
+      br()
+    ),
+    # * INTRA-CLASS CORRELATION ####
+    # tabPanel("Intra-class correlation",
+    #   value = "ICC",
+    #   h3("Intra-class correlation"),
+    #   p("More generally, Cronbach's \\(\\alpha\\) is equivalent to intraclass correlation under one of the ANOVA models."),
+    #   tableOutput("reliability_icc_table"), # preview, under construction
+    #   h4("Selected R code"),
+    #   div(code(HTML("library(psychometric)<br>library(ShinyItemAnalysis)<br><br>#&nbsp;loading&nbsp;data<br>data(HCI)<br>data&nbsp;<-&nbsp;HCI[,&nbsp;1:20]<br><br>#&nbsp;Cronbach's&nbsp;alpha&nbsp;with&nbsp;confidence&nbsp;interval<br>a&nbsp;<-&nbsp;psychometric::alpha(data)<br>psychometric::alpha.CI(a,&nbsp;N&nbsp;=&nbsp;nrow(data),&nbsp;k&nbsp;=&nbsp;ncol(data),&nbsp;level&nbsp;=&nbsp;0.95)"))),
+    #   br()
+    # ),
+
+    # * RANGE-RESTRICTED RELIABILITY ####
+    tabPanel("Restricted range",
+      value = "rr_irr",
+      h3("Range-restricted reliability"),
+      p(
+        "This section illustrates the issue of range-restricted reliability and the difficulties with the maximum
+        likelihood estimation, described in more detail in the context of inter-rater reliability in grant proposal review in ",
+        a("Erosheva, Martinkova & Lee (2021)",
+          href = "https://rss.onlinelibrary.wiley.com/loi/1467985x",
+          target = "_blank", .noWS = "outside"
+        ),
+        ". To replicate their examples, select the ", code("AIBS"), "toy dataset in the ", strong("Data"), "section."
+      ),
+      p(
+        "Below, you may select the ratio and type of range restriction given by the ", strong("proportion of ratees"),
+        ". A ratee is the object or subject that is being evaluated. It could be a proposal application in grant review
+        (as is the case in the ", code("AIBS"), "dataset), a student in educational assessment,
+        a job application in hiring, a patient in a medical study, etc.
+        Further, you may select the ", strong("direction"), "of restriction (top or bottom).
+        The left plot illustrates the variability in ratings for the whole dataset outshading the data which would be lost
+        due to range-restriction. The right plot provides the estimates of the calculated inter-rater reliability estimates,
+        defined by intraclass corelation in the simplest model including the ratee effect only.
+        The ", strong("number of bootstrap samples"), "is used to calculate the 95% confidence interval of the
+        restricted-range reliability estimate plotted in the right figure."
+      ),
+      fluidRow(
+        column(
+          2,
+          sliderInput(
+            inputId = "reliability_restricted_proportion",
+            label = "Proportion of ratees",
+            min = 0,
+            max = 100,
+            step = 1,
+            value = 100, animate = animationOptions(2000), post = "%"
+          )
+        ),
+        column(
+          2,
+          selectInput(
+            inputId = "reliability_restricted_direction",
+            label = "Direction",
+            choices = c("top", "bottom"),
+            selected = "top"
+          )
+        ),
+        column(
+          2,
+          numericInput(
+            inputId = "reliability_restricted_bootsamples",
+            label = "Num. of bootstrap samples",
+            value = 10,
+            min = 3,
+            max = 1000
+          )
+        ),
+        column(
+          2,
+          actionButton(
+            inputId = "reliability_restricted_clear",
+            label = "Clear everything",
+            icon = icon("eraser"),
+            style = "margin-top:25px"
+          )
+        )
+      ),
+      fluidRow(
+        column(6, plotlyOutput("reliability_restricted_caterpillarplot")),
+        column(6, plotlyOutput("reliability_restricted_iccplot"))
+      ),
+      # tableOutput("res_entries"),
+      br(),
+
+      # plotlyOutput("irr3_plt"),
+      textOutput("icc_text"),
+      h4("Selected R code"),
+      code(HTML("library(ShinyItemAnalysis)<br><br>#&nbsp;estimate&nbsp;inter-rater&nbsp;reliability&nbsp;(ICC)&nbsp;for&nbsp;complete&nbsp;AIBS&nbsp;dataset<br>ICCrestricted(Data&nbsp;=&nbsp;AIBS,&nbsp;case&nbsp;=&nbsp;\"ID\",&nbsp;var&nbsp;=&nbsp;\"Score\",&nbsp;rank&nbsp;=&nbsp;\"ScoreRankAdj\")<br><br>#&nbsp;estimate&nbsp;range-restricted&nbsp;ICC<br>ICCrestricted(<br>&nbsp;&nbsp;Data&nbsp;=&nbsp;AIBS,&nbsp;case&nbsp;=&nbsp;\"ID\",&nbsp;var&nbsp;=&nbsp;\"Score\",&nbsp;rank&nbsp;=&nbsp;\"ScoreRankAdj\",<br>&nbsp;&nbsp;sel&nbsp;=&nbsp;0.8,&nbsp;dir&nbsp;=&nbsp;\"bottom\"<br>)<br><br>#&nbsp;estimate&nbsp;all&nbsp;possible&nbsp;top-restricted&nbsp;subsets,&nbsp;save&nbsp;them&nbsp;to&nbsp;a&nbsp;tibble<br>all_top_restricted&nbsp;<-&nbsp;purrr::map_dfr(<br>&nbsp;&nbsp;2:72,<br>&nbsp;&nbsp;~&nbsp;ICCrestricted(<br>&nbsp;&nbsp;&nbsp;&nbsp;Data&nbsp;=&nbsp;AIBS,&nbsp;case&nbsp;=&nbsp;\"ID\",&nbsp;var&nbsp;=&nbsp;\"Score\",&nbsp;rank&nbsp;=&nbsp;\"ScoreRankAdj\",<br>&nbsp;&nbsp;&nbsp;&nbsp;sel&nbsp;=&nbsp;.x,&nbsp;nsim&nbsp;=&nbsp;10<br>&nbsp;&nbsp;)<br>)<br><br>#&nbsp;plot<br>all_top_restricted&nbsp;%>%<br>&nbsp;&nbsp;ggplot(aes(prop_sel,&nbsp;y&nbsp;=&nbsp;ICC1,&nbsp;ymin&nbsp;=&nbsp;ICC1_LCI,&nbsp;ymax&nbsp;=&nbsp;ICC1_UCI))&nbsp;+<br>&nbsp;&nbsp;geom_pointrange()&nbsp;+<br>&nbsp;&nbsp;scale_x_continuous(labels&nbsp;=&nbsp;scales::percent)&nbsp;+<br>&nbsp;&nbsp;coord_cartesian(ylim&nbsp;=&nbsp;c(0,&nbsp;1))")),
       br()
     )
   )